Save image size in ImageGradient struct to avoid recalculating
Every function dealing with an ImageGradients data structure previously had to
evaluate the size of the image by using len() on the slices. While this is not a
particularly expensive operation, there is no point in repeating it many times
when it could just be directly stored in the ImageGradient struct when it is
created, since the values are required for creating the struct anyway. Then
functions which need to evaluate the size can simply read from the data
structure what size is being dealt with.