Fix: Nonmax suppression implementation fixed
Previous version updated the edge values in place during the run. While fast,
this destroys the ability to look backwards to compare trailing edges to a
previous non-retained, yet higher value. Now, we instead store the values of
pixels to be retained and discarded, and to the suppression in a separate
operation.